move employee to separate controller
