+++ WebStore/Startup.cs
@@ -18,7 +18,7 @@ namespace WebStore
         public void ConfigureServices(IServiceCollection services)
         {
             //add MVC infrastructure 
-            services.AddControllersWithViews();
+            services.AddControllersWithViews().AddRazorRuntimeCompilation();
         }
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
@@ -26,6 +26,7 @@ namespace WebStore
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
+                app.UseBrowserLink();
             }
 
             //use static files from wwwroot
