+++ WebStore/Views/Employees/EmployeeDetails.cshtml
@@ -1,28 +1,28 @@
 ﻿@model WebStore.Model.Employee
 
     <div class="container">
-        <table>
+        <table class="table">
             <tr>
-                <td>Фамилия:</td>
+                <td>Фамилия</td>
                 <td>@Model.Surname</td>
             </tr>
             <tr>
-                <td>Имя:</td>
+                <td>Имя</td>
                 <td>@Model.Name</td>
             </tr>
             <tr>
-                <td>Отчество:</td>
+                <td>Отчество</td>
                 <td>@Model.Patronymic</td>
             </tr>
             <tr>
-                <td>Дата рождения:</td>
+                <td>Дата рождения</td>
                 <td>@Model.BirthDateTime.ToShortDateString()</td>
             </tr>
             <tr>
-                <td>Возраст:</td>
+                <td>Возраст</td>
                 <td>@Model.Age</td>
             </tr>
         </table>
 
-        <a href="/Home/Index">К списку сотрудников</a>
+        <a asp-controller="Employees" asp-action="Index">К списку сотрудников</a>
     </div>
